Make libedit usage optional in Unix REPL
For the simple Unix REPL program (source/main.cpp), make usage of libedit
optional for history and interactive input. The alternative falls-back to
using the getline() function available in GNU libc and in the libc of most
BSD systems.

Also, the Makefile now allows for easily defining build options, for
example, to build without any option:

  % make DEFINES=''

And, to force usage of libedit:

  % make DEFINES='OS_USE_LIBEDIT'